refactor(pci.projects.failover-ips): remove reference to Agora (#816)
ref: MANAGER-2753
