+++ packages/manager/modules/pci/src/projects/project/failover-ips/onboarding/order/order.routing.js
@@ -36,7 +36,7 @@ export default /* @ngInject */ ($stateProvider, coreConfigProvider) => {
         layout: 'modal',
         views: {
           modal: {
-            component: 'pciProjectFailoverIpsAgoraOrder',
+            component: 'pciProjectFailoverIpsOrderOrder',
           },
         },
         resolve: {
