+++ packages/manager/modules/pci/src/projects/project/failover-ips/order/order/index.js
@@ -4,14 +4,14 @@ import 'ovh-api-services';
 
 import component from './order.component';
 
-const moduleName = 'ovhManagerPciProjectFailoverIpsAgoraOrder';
+const moduleName = 'ovhManagerPciProjectFailoverIpsOrderOrder';
 
 angular
   .module(moduleName, [
     'pascalprecht.translate',
     'ovh-api-services',
   ])
-  .component('pciProjectFailoverIpsAgoraOrder', component)
+  .component('pciProjectFailoverIpsOrderOrder', component)
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
