+++ packages/manager/modules/pci/src/projects/project/failover-ips/order/order/translations/Messages_fr_FR.json
